Rydder litt mer i setup-lokal-utvikling.sh

+++ lokal-utvikling/setup-lokal-utvikling.sh
@@ -120,22 +120,24 @@ fi
 mkdir $folder
 
 
+# Copy all docker-compose file to separate docker-compose-lokal folder for local instance
 cd ../resources/pipeline
 settPorterSomSkalErstattes ${applikasjoner[0]}
 for f in {.*,*}; do
     if [[ -f "$f" ]]; then
-      cp "$f" "${relativ_path}/$f"
+        cp "$f" "${relativ_path}/$f"
     elif [[ $f == oracle-init ]] || [[ $f == postgres-init ]] || [[ $f == tokenx ]]; then
-      cp -r "$f" "${relativ_path}"
+        cp -r "$f" "${relativ_path}"
     fi
 done
 
-
+# Redirect alle kall for angitte applikasjoner lokalt på maskinen din. F.eks. ved valg av fpabonnent erstattes
+# http://fpabonnent:8080/fpabonnent med http://host.docker.internal:8065/fpabonnent
 cd $relativ_path
 for applikasjon in "${applikasjoner[@]}"; do
   settPorterSomSkalErstattes $applikasjon
   for f in {.*,*}; do
-    if [ -f "$f" ] && [[ $f != .env ]]  && [[ $f != *.sh ]]; then
+    if [ -f "$f" ] && [[ $f != .env ]]  && [[ $f != *.sh ]] && [[ $f != *.sql ]]; then # Only files && not .env %% not script files && not sql files
       for ((i=0;i<${#replace_port_array[@]};++i)); do
         sed -i.bak "s/${applikasjon}:${replace_port_array[i]}/${host_adresse}:${with_port_array[i]}/g" "$f"
         rm $f.bak
@@ -144,27 +146,30 @@ for applikasjon in "${applikasjoner[@]}"; do
   done
 done
 
-applikasjoner_som_kan_mockes_ut=(fpoppdrag fptilbake fpformidling fprisk)
+# Ved mocking av våre interne applikasjoner bruker vi en stub i vtp. Bare støtte for fpoppdrag, fptilbake, fpformidling eller fprisk
 if [ ${#mock_applikasjon[@]} ]; then
     echo "Mocker ut følgende applikasjoner: ${mock_applikasjon[@]}"
+    applikasjoner_som_kan_mockes_ut=(fpoppdrag fptilbake fpformidling fprisk)
     for applikasjon in "${mock_applikasjon[@]}"; do
-      if [[ "${applikasjoner_som_kan_mockes_ut[@]}" =~ "${applikasjon}" ]]; then
-        if [[ "$*" == *vtp* ]]; then
-          sed -i.bak "s*${applikasjon}:8080*${host_adresse}:8060/rest/dummy*g" "docker-compose.yml"
-          sed -i.bak "s*${applikasjon}:8080*${host_adresse}:8060/rest/dummy*g" "common.env"
+        if [[ "${applikasjoner_som_kan_mockes_ut[@]}" =~ "${applikasjon}" ]]; then
+            if [[ "$*" == *vtp* ]]; then
+                sed -i.bak "s*${applikasjon}:8080*${host_adresse}:8060/rest/dummy*g" "docker-compose.yml"
+                sed -i.bak "s*${applikasjon}:8080*${host_adresse}:8060/rest/dummy*g" "common.env"
+            else
+                sed -i.bak "s*${applikasjon}:8080*vtp:8060/rest/dummy*g" "docker-compose.yml"
+                sed -i.bak "s*${applikasjon}:8080*vtp:8060/rest/dummy*g" "common.env"
+            fi
+
+            if [[ "${applikasjon}" =~ "fpoppdrag"  ]] && [[ "$*" == *fpfrontend* ]] ; then
+                sed -i.bak "s*localhost:9000/fpoppdrag/api*vtp:8060/rest/dummy/fpoppdrag/api*g" "docker-compose.yml"
+                sed -i.bak "s*localhost:9000/fpoppdrag/api*vtp:8060/rest/dummy/fpoppdrag/api*g" "common.env"
+            fi
+
+            rm docker-compose.yml.bak
+            rm common.env.bak
         else
-          sed -i.bak "s*${applikasjon}:8080*vtp:8060/rest/dummy*g" "docker-compose.yml"
-          sed -i.bak "s*${applikasjon}:8080*vtp:8060/rest/dummy*g" "common.env"
+            echo "Mock av ${applikasjon} finnes ikke – beholder konfigurasjon som den er"
         fi
-        if [[ "${applikasjon}" =~ "fpoppdrag"  ]] && [[ "$*" == *fpfrontend* ]] ; then
-          sed -i.bak "s*localhost:9000/fpoppdrag/api*vtp:8060/rest/dummy/fpoppdrag/api*g" "docker-compose.yml"
-          sed -i.bak "s*localhost:9000/fpoppdrag/api*vtp:8060/rest/dummy/fpoppdrag/api*g" "common.env"
-        fi
-        rm docker-compose.yml.bak
-        rm common.env.bak
-      else
-        echo "Mock av ${applikasjon} finnes ikke – beholder konfigurasjon som den er."
-      fi
     done
 fi
